### 2025-01-09 #232 @T-Strojny
BlockingPy submission
Submitting Author: (@T-Strojny)
All current maintainers: @T-Strojny
Package Name: BlockingPy
One-Line Description of Package: Blocking records for record linkage and deduplication with Approximate Nearest Neighbor algorithms.;
Repository Link: https://github.com/ncn-foreigners/BlockingPy
Version submitted: v0.1.7
EiC: TBD
Version accepted: TBD

---

- Include a brief paragraph describing what your package does: BlockingPy is a package that speeds up record linkage and deduplication tasks by using  Approximate Nearest Neighbor (ANN) algorithms to create blocks with candidate record pairs. When linking or deduplicating large datasets, comparing all possible record pairs becomes computationally infeasible. BlockingPy solves this by using ANN algorithms to quickly identify similar records while significantly reducing the number of required comparisons.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

  - Data processing/munging : BlockingPy transforms raw data into feature vectors and applies ANN algorithms  and graphs to reduce the comparison space which enables scalable record linkage and deduplication.

  BlockingPy is targeted for data scientists, researchers, and analysts working with large datasets that require record matching or deduplication and need a scalable approach.
There are many packages around Record Linkage, however ours specializes in the blocking task and uses novel approach which is the use of ANN algorithms.
No inquiry was made
