### 2025-01-17 35ff086
added author, funding info
