+++ pyproject.toml
@@ -2,10 +2,10 @@
 name = "blockingpy"
 version = "0.1.7"
 description = "Blocking records for record linkage and data deduplication based on ANN algorithms."
-authors = ["Tymoteusz Strojny <tymek.strojny@gmail.com>"]
+authors = ["Tymoteusz Strojny <tymek.strojny@gmail.com>", "Maciej Beręsewicz <maciej.beresewicz@ue.poznan.pl>"]
 readme = "README.md"
 license = "MIT"
-repository = "https://github.com/T-Strojny/BlockingPy"
+repository = "https://github.com/ncn-foreigners/BlockingPy"
 documentation = "https://blockingpy.readthedocs.io/en/latest/"
 keywords = ["record-linkage", "deduplication", "ANN", "blocking", "data-matching"]
 packages = [{include = "blockingpy"}]
@@ -17,12 +17,14 @@ classifiers = [
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Scientific/Engineering :: Information Analysis",
     "Topic :: Software Development :: Libraries :: Python Modules"
 ]
 
 [tool.poetry.urls]
-"Bug Tracker" = "https://github.com/T-Strojny/BlockingPy/issues"
+"Bug Tracker" = "https://github.com/ncn-foreigners/BlockingPy/issues"
+"Funding" = "https://www.ncn.gov.pl"
 
 [tool.poetry.dependencies]
 python = "^3.10"
