+++ docs/examples/record_linkage.md
@@ -1,12 +1,16 @@
 (record_linkage)=
 # Record Linkage
 
-This example demonstrates how to use BlockingPy for record linkage between two datasets. We'll use example data from the URos 2021 Conference tutorial which contains:
+This example demonstrates how to use BlockingPy for record linkage between two datasets. We'll use example data created by Paula McLeod, Dick Heasman and Ian Forbes, ONS,
+    for the ESSnet DI on-the-job training course, Southampton,
+    25-28 January 2011:
 
 - Census: A fictional dataset representing observations from a decennial Census
 - CIS: Fictional observations from Customer Information System (combined administrative data from tax and benefit systems)
 
 Some records in the CIS dataset contain Census person IDs, which we'll use to evaluate our blocking performance.
+
+This datasets come with the `BlockingPy` package and can be accesed via `load_census_cis_data` function from `blockingpy.datasets`.
 
 ## Setup
 
@@ -20,6 +24,7 @@ Import required packages:
 
 ```python
 from blockingpy import Blocker
+from blockingpy.datasets import load_census_cis_data
 import pandas as pd
 ```
 
@@ -28,8 +33,14 @@ import pandas as pd
 Download example data:
 
 ```python
-census = pd.read_csv("https://raw.githubusercontent.com/djvanderlaan/tutorial-reclin-uros2021/main/data/census.csv")
-cis = pd.read_csv("https://raw.githubusercontent.com/djvanderlaan/tutorial-reclin-uros2021/main/data/cis.csv")
+census, cis = load_census_cis_data()
+```
+
+Firstly, we need to filter only those columns which we'll need:
+
+```python
+census = census[["PERSON_ID", "PERNAME1", "PERNAME2", "SEX", "DOB_DAY", "DOB_MON", "DOB_YEAR", "ENUMCAP", "ENUMPC"]]
+cis = cis[["PERSON_ID", "PERNAME1", "PERNAME2", "SEX", "DOB_DAY", "DOB_MON", "DOB_YEAR", "ENUMCAP", "ENUMPC"]]
 ```
 
 Let's take a look at the data:
@@ -37,14 +48,14 @@ Let's take a look at the data:
 ```python
 print(census.head())
 
-#       person_id pername1 pername2 sex  dob_day  dob_mon  dob_year  \
+#       PERSON_ID PERNAME1 PERNAME2 SEX  DOB_DAY  DOB_MON  DOB_YEAR  \
 # 0  DE03US001001    COUIE    PRICE   M      1.0        6    1960.0   
 # 1  DE03US001002    ABBIE    PVICE   F      9.0       11    1961.0   
 # 2  DE03US001003    LACEY    PRICE   F      7.0        2    1999.0   
 # 3  DE03US001004   SAMUEL    PRICE   M     13.0        4    1990.0   
 # 4  DE03US001005   JOSEPH    PRICE   M     20.0        4    1986.0   
 
-#           enumcap  enumpc  
+#           ENUMCAP  ENUMPC  
 # 0  1 WINDSOR ROAD  DE03US  
 # 1  1 WINDSOR ROAD  DE03US  
 # 2  1 WINDSOR ROAD  DE03US  
@@ -53,14 +64,14 @@ print(census.head())
 
 print(cis.head())
 
-#   person_id  pername1  pername2 sex  dob_day  dob_mon  dob_year  \
+#   PERSON_ID  PERNAME1  PERNAME2 SEX  DOB_DAY  DOB_MON  DOB_YEAR  \
 # 0       NaN    HAYDEN      HALL   M      NaN        1       NaN   
[…]
