+++ blockingpy/__init__.py
@@ -5,7 +5,11 @@ import sys
 
 from .blocker import Blocker
 
-__all__ = ["Blocker"]
+__all__ = [
+    'Blocker',
+    'load_census_cis_data',
+    'load_deduplication_data'
+]
 
 logger = logging.getLogger("blockingpy")
 handler = logging.StreamHandler(sys.stdout)
