+++ README.md
@@ -147,6 +147,17 @@ print(dedup_result.result)
 
 You can find detailed information about BlockingPy in [documentation](https://blockingpy.readthedocs.io/en/latest/).
 
+## Example Datasets
+
+BlockingPy comes with built-in example datasets:
+
+- Census-Cis dataset created by Paula McLeod, Dick Heasman and Ian Forbes, ONS,
+    for the ESSnet DI on-the-job training course, Southampton,
+    25-28 January 2011
+
+- Deduplication dataset taken from [RecordLinkage](https://cran.r-project.org/package=RecordLinkage) R package developed by Murat Sariyar
+    and Andreas Borg. Package is licensed under GPL-3 license. Also known as [RLdata10000](https://www.rdocumentation.org/packages/RecordLinkage/versions/0.4-12.4/topics/RLdata).
+
 ## Disclaimer
 BlockingPy is still under development, API and features may change. Also bugs or errors can occur. 
 
