### 2025-01-17 5c63f42
bump version to 0.1.8, include data in package
