### 2025-01-21 #232 @coatless
---
---

BlockingPy is in pristine shape for moving forward with a review! Nice work on getting it packaged for Python and implemented. Happy to see both `mlpack` and the original note on the R blocking package being emphasized.
