### 2025-01-25 251cd3b
update docs after batch size addition
