### 2025-01-25 7c48f97
added batch size processing for evaluation
