+++ blockingpy/blocker.py
@@ -155,6 +155,9 @@ class Blocker:
         2. Sparse matrices (scipy.sparse.csr_matrix) as a Document-Term Matrix (DTM)
         3. Dense matrices (numpy.ndarray) as a Document-Term Matrix (DTM)
 
+        For evaluation of larger datasets, we recommend using the separate eval() method
+            since it allows you to set the batch size for evaluation.
+
         For text data, additional preprocessing is performed using
         the parameters in control_txt.
 
@@ -171,6 +174,7 @@ class Blocker:
         self.y_colnames = y_colnames
         self.control_ann = controls_ann(control_ann)
         self.control_txt = controls_txt(control_txt)
+        self.EVAL_BATCH_SIZE = 10_000
 
         if deduplication:
             self.y_colnames = self.x_colnames
@@ -266,7 +270,6 @@ class Blocker:
             x_df["query_g"] = "q" + x_df["y"].astype(str)
             x_df["index_g"] = "i" + x_df["x"].astype(str)
 
-        # IGRAPH PART IN R
         x_gr = nx.from_pandas_edgelist(
             x_df, source="query_g", target="index_g", create_using=nx.Graph()
         )
@@ -288,37 +291,24 @@ class Blocker:
                 sorted_dict[key] = x_block[key]
 
         x_df["block"] = x_df["query_g"].apply(lambda x: x_block[x] if x in x_block else None)
-        ###
 
         if true_blocks is not None:
             if not deduplication:
-                candidate_pairs = list(itertools.product(list(range(len(x_dtm))), true_blocks["y"]))
-                cp_df = pd.DataFrame(candidate_pairs, columns=["x", "y"])
-                cp_df = cp_df.astype(int)
-                comparison_df = (
-                    cp_df.merge(true_blocks, on=["x", "y"], how="left")
-                    .rename(columns={"block": "block_true"})
-                    .merge(x_df, on=["x", "y"], how="left")
-                    .rename(columns={"block": "block_pred"})
-                )
-                comparison_df["TP"] = (comparison_df["block_true"].notna()) & (
-                    comparison_df["block_pred"].notna()
-                )
-                # CNL -> Correct Non-Links / True Negative
-                comparison_df["CNL"] = (comparison_df["block_true"].isna()) & (
-                    comparison_df["block_pred"].isna()
-                )
-                comparison_df["FP"] = (comparison_df["block_true"].isna()) & (
-                    comparison_df["block_pred"].notna()
-                )
-                comparison_df["FN"] = (comparison_df["block_true"].notna()) & (
-                    comparison_df["block_pred"].isna()
-                )
+                total_tn = total_fp = total_fn = total_tp = 0
+                batch_size = 10_000
+                n_records = len(x_dtm)
+
+                for start_idx in range(0, n_records, batch_size):
+                    tn, fp, fn, tp = self._process_record_linkage_batch(
+                        start_idx, batch_size, n_records, true_blocks, x_df
+                    )
+                    total_tn += tn
+                    total_fp += fp
+                    total_fn += fn
+                    total_tp += tp
+
                 self.confusion = pd.DataFrame(
-                    [
-                        [comparison_df["CNL"].sum(), comparison_df["FN"].sum()],
-                        [comparison_df["FP"].sum(), comparison_df["TP"].sum()],
-                    ],
+                    [[total_tn, total_fn], [total_fp, total_tp]],
                     index=["Predicted Negative", "Predicted Positive"],
                     columns=["Actual Negative", "Actual Positive"],
                 )
@@ -332,37 +322,38 @@ class Blocker:
                     .rename(columns={"block": "true_id"})
[…]
