+++ blockingpy/datasets/base.py
@@ -1,6 +1,5 @@
 """Functions to load built-in BlockingPy datasets."""
 
-
 import pandas as pd
 
 from .utils import get_data_file
