+++ blockingpy/blocker.py
@@ -156,7 +156,7 @@ class Blocker:
         3. Dense matrices (numpy.ndarray) as a Document-Term Matrix (DTM)
 
         For evaluation of larger datasets, we recommend using the separate eval() method
-            since it allows you to set the batch size for evaluation.
+        since it allows you to set the batch size for evaluation.
 
         For text data, additional preprocessing is performed using
         the parameters in control_txt.
@@ -174,7 +174,6 @@ class Blocker:
         self.y_colnames = y_colnames
         self.control_ann = controls_ann(control_ann)
         self.control_txt = controls_txt(control_txt)
-        self.EVAL_BATCH_SIZE = 10_000
 
         if deduplication:
             self.y_colnames = self.x_colnames
@@ -203,9 +202,11 @@ class Blocker:
         if y is not None:
             deduplication = False
             k = 1
+            len_y = y.shape[0]
         else:
             y = x
             k = 2
+            len_y = None
 
         InputValidator.validate_true_blocks(true_blocks, deduplication)
 
@@ -257,7 +258,7 @@ class Blocker:
             verbose=True if verbose in {2, 3} else False,
             controls=self.control_ann,
         )
-        logger.info("===== creating graph =====\n")
+        logger.info("===== creating graph =====")
 
         if deduplication:
             x_df["pair"] = x_df.apply(lambda row: tuple(sorted([row["y"], row["x"]])), axis=1)
@@ -293,53 +294,55 @@ class Blocker:
         x_df["block"] = x_df["query_g"].apply(lambda x: x_block[x] if x in x_block else None)
 
         if true_blocks is not None:
+            logger.info("===== evaluating =====")
+            total_tn = total_fp = total_fn = total_tp = 0
+            batch_size = 1000
+
             if not deduplication:
-                total_tn = total_fp = total_fn = total_tp = 0
-                batch_size = 10_000
-                n_records = len(x_dtm)
+                unique_tb_x = true_blocks["x"].unique()
+                unique_tb_y = true_blocks["y"].unique()
 
-                for start_idx in range(0, n_records, batch_size):
-                    tn, fp, fn, tp = self._process_record_linkage_batch(
-                        start_idx, batch_size, n_records, true_blocks, x_df
-                    )
-                    total_tn += tn
-                    total_fp += fp
-                    total_fn += fn
-                    total_tp += tp
+                for start_idx_x in range(0, len(unique_tb_x), batch_size):
+                    sub_x = unique_tb_x[start_idx_x : start_idx_x + batch_size]
+                    for start_idx_y in range(0, len(unique_tb_y), batch_size):
+                        sub_y = unique_tb_y[start_idx_y : start_idx_y + batch_size]
 
-                self.confusion = pd.DataFrame(
-                    [[total_tn, total_fn], [total_fp, total_tp]],
-                    index=["Predicted Negative", "Predicted Positive"],
-                    columns=["Actual Negative", "Actual Positive"],
-                )
+                        tp, fp, fn = self._eval_rl_batch(sub_x, sub_y, true_blocks, x_df)
+
+                        total_tp += tp
+                        total_fp += fp
+                        total_fn += fn
+                total_tn = len(unique_tb_x) * len(unique_tb_y) - total_tp - total_fp - total_fn
 
             else:
-                pairs_to_eval_long = (
-                    pd.melt(x_df[["x", "y", "block"]], id_vars=["block"])[["block", "value"]]
-                    .drop_duplicates()
-                    .rename(columns={"block": "block_id", "value": "x"})
-                    .merge(true_blocks[["x", "block"]], on="x", how="left")
[…]
