+++ docs/changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.1.10
+- evaluation only for records that exist in true blocks.
+- default distance for `faiss` changed to `cosine`
+- code simplification
+- minor changes
+
 ## v0.1.9
 - optimized evaluation part to allow batch processing
 
