+++ blockingpy/blocker.py
@@ -302,12 +302,31 @@ class Blocker:
                 unique_tb_x = true_blocks["x"].unique()
                 unique_tb_y = true_blocks["y"].unique()
 
+                true_x_blocks = true_blocks[["x", "block"]].drop_duplicates()
+                true_y_blocks = true_blocks[["y", "block"]].drop_duplicates()
+                pred_x_blocks = x_df[["x", "block"]].drop_duplicates()
+                pred_y_blocks = x_df[["y", "block"]].drop_duplicates()
+
+                total_batches_x = (len(unique_tb_x) + batch_size - 1) // batch_size
+                total_batches_y = (len(unique_tb_y) + batch_size - 1) // batch_size
+                total_batches = total_batches_x * total_batches_y
+
                 for start_idx_x in range(0, len(unique_tb_x), batch_size):
+
+                    current_batch_x = (start_idx_x // batch_size) + 1
+
                     sub_x = unique_tb_x[start_idx_x : start_idx_x + batch_size]
+
                     for start_idx_y in range(0, len(unique_tb_y), batch_size):
+                        current_batch_y = (start_idx_y // batch_size) + 1
+                        current_batch = ((current_batch_x - 1) * total_batches_y) + current_batch_y
+                        logger.info(f"Evaluating batch {current_batch} of {total_batches}")
+
                         sub_y = unique_tb_y[start_idx_y : start_idx_y + batch_size]
 
-                        tp, fp, fn = self._eval_rl_batch(sub_x, sub_y, true_blocks, x_df)
+                        tp, fp, fn = self._eval_rl_batch(
+                            sub_x, sub_y, true_x_blocks, true_y_blocks, pred_x_blocks, pred_y_blocks
+                        )
 
                         total_tp += tp
                         total_fp += fp
@@ -321,9 +340,18 @@ class Blocker:
                     .rename(columns={"x_x": "x"})
                 )
                 unique_tb_x = true_blocks["x"].unique()
+
+                total_batches_x = (len(unique_tb_x) + batch_size - 1) // batch_size
+                total_batches = total_batches_x * total_batches_x
+
                 for start_idx in range(0, len(unique_tb_x), batch_size):
+                    current_batch_x = (start_idx // batch_size) + 1
                     sub_x = unique_tb_x[start_idx : start_idx + batch_size]
                     for start_idx_y in range(0, len(unique_tb_x), batch_size):
+                        current_batch_y = (start_idx_y // batch_size) + 1
+                        current_batch = ((current_batch_x - 1) * total_batches_x) + current_batch_y
+                        logger.info(f"Evaluating batch {current_batch} of {total_batches}")
+
                         sub_y = unique_tb_x[start_idx_y : start_idx_y + batch_size]
 
                         tp, fp, fn = self._eval_dedup_batch(sub_x, sub_y, true_blocks, x_df_long)
@@ -438,13 +466,27 @@ class Blocker:
             unique_tb_x = true_blocks["x"].unique()
             unique_tb_y = true_blocks["y"].unique()
 
+            true_x_blocks = true_blocks[["x", "block"]].drop_duplicates()
+            true_y_blocks = true_blocks[["y", "block"]].drop_duplicates()
+            pred_x_blocks = blocking_result.result[["x", "block"]].drop_duplicates()
+            pred_y_blocks = blocking_result.result[["y", "block"]].drop_duplicates()
+
+            total_batches_x = (len(unique_tb_x) + batch_size - 1) // batch_size
+            total_batches_y = (len(unique_tb_y) + batch_size - 1) // batch_size
+            total_batches = total_batches_x * total_batches_y
+
[…]
