+++ docs/examples/record_linkage.md
@@ -89,8 +89,8 @@ Preprocess data and create column `txt` containing concatenated variables:
 
 ```python
 # Convert numeric fields to strings
-census[['DOB_DAY', 'DOB_MON', 'DOB_YEAR']] = census[['DOB_DAY', 'DOB_MON', 'DOB_YEAR']].astype(str)
-cis[['DOB_DAY', 'DOB_MON', 'DOB_YEAR']] = cis[['DOB_DAY', 'DOB_MON', 'DOB_YEAR']].astype(str)
+census[['DOB_DAY', 'DOB_MON', 'DOB_YEAR']] = census[['DOB_DAY', 'DOB_MON', 'DOB_YEAR']].astype("Int64").astype(str).replace('<NA>', '')
+cis[['DOB_DAY', 'DOB_MON', 'DOB_YEAR']] = cis[['DOB_DAY', 'DOB_MON', 'DOB_YEAR']].astype("Int64").astype(str).replace('<NA>', '')
 
 # Fill NAs with empty strings
 census = census.fillna('')
@@ -112,20 +112,20 @@ Let's see how the new column looks like:
 print(census['txt'].head())
 
 # txt
-# 0	COUIEPRICEM1.061960.01 WINDSOR ROADDE03US
-# 1	ABBIEPVICEF9.0111961.01 WINDSOR ROADDE03US
-# 2	LACEYPRICEF7.021999.01 WINDSOR ROADDE03US
-# 3	SAMUELPRICEM13.041990.01 WINDSOR ROADDE03US
-# 4	JOSEPHPRICEM20.041986.01 WINDSOR ROADDE03US
+# 0      COUIEPRICEM1619601 WINDSOR ROADDE03US
+# 1     ABBIEPVICEF91119611 WINDSOR ROADDE03US
+# 2      LACEYPRICEF7219991 WINDSOR ROADDE03US
+# 3    SAMUELPRICEM13419901 WINDSOR ROADDE03US
+# 4    JOSEPHPRICEM20419861 WINDSOR ROADDE03US
 
 print(cis['txt'].head())
 
 # 	txt
-# 0	HAYDENHALLMnan1nan91 CLARENCE ROADPO827ER
-# 1	SERENANDERSONF1.01nan24 CHURCH LANELS992DB
-# 2	LEWISLEWISM1.01nan53 CHURCH ROADM432ZZ
-# 3	HARRISONPOSTERM5.01nan19 HIGHFIELD ROADSW75TG
-# 4	MUHAMMEDWATSUNM7.01nan17 VICTORIA STREET
+# 0         HAYDENHALLM191 CLARENCE ROADPO827ER
+# 1       SERENANDERSONF1124 CHURCH LANELS992DB
+# 2           LEWISLEWISM1153 CHURCH ROADM432ZZ
+# 3    HARRISONPOSTERM5119 HIGHFIELD ROADSW75TG
+# 4         MUHAMMEDWATSUNM7117 VICTORIA STREET
 
 ```
 
@@ -174,46 +174,47 @@ print(rec_lin_result)
 # ========================================================
 # Distribution of the size of the blocks:
 # Block Size | Number of Blocks
-#          2 | 23392          
-#          3 | 592            
-#          4 | 11             
-#          5 | 1   
+        #  2 | 23390          
+        #  3 | 590            
+        #  4 | 13             
+        #  5 | 1    
 
 print(rec_lin_result.result.head())
 
-#            x      y  block      dist
-# 0      17339      0      0  0.273628
-# 1       9567      1      1  0.103388
-# 2      10389      2      2  0.107852
-# 3      24258      3      3  0.211039
-# 4       3714      4      4  0.294986
+#      x      y  block      dist
[…]
