### 2025-02-18 c98e414
added note to rl example [skip ci]
