+++ blockingpy/faiss_blocker.py
@@ -161,6 +161,12 @@ class FaissBlocker(BlockingMethod):
             )
 
         distances, indices = self.index.search(x=y, k=k_search)
+
+        if distance == "cosine":
+            # faiss returns cosine similarity ([0,1] since there are only non-negative values in DTM),
+             # we want cosine distance
+            distances = 1 - distances 
+
         if k == 2:
             indices, distances = rearrange_array(indices, distances)
 
