### 2025-02-20 931c1d3
new example in docs
