+++ docs/examples/index.md
@@ -7,3 +7,4 @@
 record_linkage
 deduplication
 deduplication_2
+recordlinkage_integration
