+++ docs/examples/recordlinkage_integration.md
@@ -79,76 +79,52 @@ print(res.result.head())
 
 ## Integration
 
-To integrate our results, we need to obtain `pd.MultiIndex` with indexes of desired indicies in each pair. For that we need to create additional variables:
+To integrate our results, we can add a `block` column to the original dataframe, which we can do by melting the blocking result and merging it with the original dataframe.
 
 ```python
-df['x'] = range(len(df))
-df['id'] = df.index
-```
-Now, we can merge the blocking result with main dataframe to extract information about indexes in each block:
+result_df = blocking_result.result
 
-```python
-df_merged = res.result.merge(
-    df[['x', 'id']],
-    on='x',
-    how='left'
+mapping_df = (
+    result_df
+    .melt(id_vars=['block'], value_vars=['x', 'y']  value_name='record_id')
+    .drop_duplicates(subset=['record_id'])
 )
-df_merged = df_merged.rename({'id':'id_x'})
-df_merged = df_merged.merge(
-    df[['x', 'id']],
-    left_on='y',
-    right_on='x',
-    how='left'
-).rename({'id':'id_y', 'x_x':'x'}, axis=1).drop(labels=['x_y'], axis=1)
-df_merged.head()
 
+record_to_block = dict(zip(mapping_df['record_id'], mapping_df['block']))
+    
+new_data = original_data.copy()
 
-# 	x	y	block	dist	        id_x        	id_y
-# 0	474	0	0	0.048376	rec-223-dup-0	rec-223-org
-# 1	330	1	1	0.038961	rec-122-dup-0	rec-122-org
-# 2	351	2	2	0.086690	rec-373-dup-0	rec-373-org
-# 3	290	3	3	0.024617	rec-10-org	rec-10-dup-0
-# 4	333	4	4	0.105662	rec-227-dup-0	rec-227-org
+new_data['block'] = [record_to_block.get(i) for i in range(len(original_data))]
+
+print(new_data[['given_name', 'block']].head(5))
+# 	        given_name block
+# rec_id		
+# rec-223-org		   0
+# rec-122-org	lachlan	   1
+# rec-373-org	deakin	   2
+# rec-10-dup-0	kayla	   3
+# rec-227-org	luke	   4
+
 ```
 
-In the next step we create `pd.MultiIndex` with candidate pairs. In deduplication tasks each record in a given block must be paired with every other record in this block:
+Now we can use the `Index` object from `recordlinkage` with the `block` column to integrate `BlockingPy` results with `recordlinkage`:
 
 ```python
-df_merged = df_merged[['id_x', 'id_y', 'block']]
-
-block_groups = (
-    df_merged.melt(id_vars=['block'], value_name='record')
-      .drop_duplicates()
-      .groupby('block')['record']
-      .agg(list)
-)
-
-all_pairs = []
-for records in block_groups:
-    all_pairs.extend(combinations(sorted(records), 2))
-
-x_indices, y_indices = zip(*all_pairs)
-
-df_end = pd.MultiIndex.from_arrays(
-    [x_indices, y_indices],
-    names=['rec_id_1', 'rec_id_2']
-)
-
-print(df_end)
-# MultiIndex([('rec-223-dup-0', 'rec-223-org'),
-#             ('rec-122-dup-0', 'rec-122-org'),
-#             ('rec-373-dup-0', 'rec-373-org'),
-#             ( 'rec-10-dup-0',  'rec-10-org'),
-#             ('rec-227-dup-0', 'rec-227-org'),
-#             (  'rec-6-dup-0',   'rec-6-org'),
+indexer = recordlinkage.Index()
+indexer.block('block')
+pairs = indexer.index(new_data)
[…]
