+++ docs/examples/splink_integration.md
@@ -1,0 +1,178 @@
+# Integration with splink package
+
+In this example, we demonstrate how to integrate `BlockingPy` with the `Splink` package for probabilistic record linkage. `Splink` provides a powerful framework for entity resolution, and `BlockingPy` can enhance its performance by providing another blocking approach.
+This example will show how to deduplicate the `fake_1000` dataset included with `Splink` using `BlockingPy` for the blocking phase improvement and `Splink` for the matching phase. We aim to follow the example available in the `Splink` documentation and modify the blocking procedure. The original can be foud [here](https://moj-analytical-services.github.io/splink/demos/examples/duckdb/accuracy_analysis_from_labels_column.html).
+
+## Setup
+First, we need to install `BlockingPy` and `Splink`:
+
+```bash
+pip install blockingpy splink
+```
+
+Import necessary components:
+
+```python
+from splink import splink_datasets, SettingsCreator, Linker, block_on, DuckDBAPI
+import splink.comparison_library as cl
+from blockingpy import Blocker
+import pandas as pd
+import numpy as np
+np.random.seed(42)
+```
+
+## Data preparation
+The `fake_1000` dataset contains 1000 records with personal information like names, dates of birth, and email addresses. The dataset consists of 251 unique entities (clusters), with each entity having one original record and various duplicates.
+
+```python
+df = splink_datasets.fake_1000
+print(df.head(5))
+#    unique_id first_name surname         dob    city                    email    cluster  
+# 0          0     Robert    Alan  1971-06-24     NaN      robert255@smith.net          0
+# 1          1     Robert   Allen  1971-05-24     NaN      roberta25@smith.net          0
+# 2          2        Rob   Allen  1971-06-24  London      roberta25@smith.net          0
+# 3          3     Robert    Alen  1971-06-24   Lonon                      NaN          0
+# 4          4      Grace     NaN  1997-04-26    Hull  grace.kelly52@jones.com          1
+```
+
+For BlockingPy, we'll create a text field combining multiple columns to allow blocking on overall record similarity:
+
+```python
+df['txt'] = df['first_name'].fillna('') + ' ' + \
+            df['surname'].fillna('') + \
+            df['dob'].fillna('') + ' ' + \
+            df['city'].fillna('') + ' ' + \
+            df['email'].fillna('')   
+```
+
+## Blocking
+
+Now we can obtain blocks from `BlockingPy`:
+
+```python
+blocker = Blocker()
+
+res = blocker.block(
+        x = df['txt'],
+        ann='faiss',
+)
+
+print(res)
+# ========================================================
+# Blocking based on the faiss method.
+# Number of blocks: 253
+# Number of columns used for blocking: 906
+# Reduction ratio: 0.996314
+# ========================================================
+# Distribution of the size of the blocks:
+# Block Size | Number of Blocks
+#          2 | 64             
+#          3 | 53             
+#          4 | 50             
+#          5 | 36             
+#          6 | 26             
+#          7 | 16             
+#          8 | 7              
+#          9 | 1   
+print(res.result.head())
+#      x  y  block      dist
+# 0    1  0      0  0.142391
+# 1    1  2      0  0.208361
+# 2    2  3      0  0.230678
+# 3    5  4      1  0.145114
+# 4  814  6      2  0.584251
+```
+
+## Results integration
+
+To integrate our results, we can add a `block` column to the original dataframe, which we can do by melting the blocking result and merging it with the original dataframe.
+
+```python
+result_df = res.result
+
+mapping_df = (
+    result_df
[…]
