+++ docs/examples/index.md
@@ -8,3 +8,4 @@ record_linkage
 deduplication
 deduplication_2
 recordlinkage_integration
+splink_integration
