### 2025-03-16 15f9fa1
improved examples
