### 2025-03-17 dc45705
added indexLSH and indexHNSWFlat from faiss
