+++ blockingpy/controls.py
@@ -169,8 +169,7 @@ def controls_ann(controls: dict[str, Any], **kwargs: Any) -> dict[str, Any]:
             "hnsw_ef_construction": 200,
             "hnsw_ef_search": 200,
             "lsh_nbits": 8,
-            "lsh_rotate_data": True
-
+            "lsh_rotate_data": True,
         },
         "algo": "lsh",
     }
