### 2025-03-19 2aa4e6d
Change default faiss lsh_nbits from 8 to 2
