+++ tests/test_blocking.py
@@ -171,8 +171,8 @@ def test_true_blocks_linkage(small_named_txt_data):
     assert all(metric in result.metrics for metric in expected_metrics)
 
     assert result.confusion.shape == (2, 2)
-    assert set(result.confusion.index) == {"Predicted Negative", "Predicted Positive"}
-    assert set(result.confusion.columns) == {"Actual Positive", "Actual Negative"}
+    assert set(result.confusion.index) == {"Actual Positive", "Actual Negative"}
+    assert set(result.confusion.columns) == {"Predicted Positive", "Predicted Negative"}
 
 
 def test_true_blocks_deduplication(small_named_txt_data):
@@ -201,8 +201,8 @@ def test_true_blocks_deduplication(small_named_txt_data):
     assert all(metric in result.metrics for metric in expected_metrics)
 
     assert result.confusion.shape == (2, 2)
-    assert set(result.confusion.index) == {"Predicted Negative", "Predicted Positive"}
-    assert set(result.confusion.columns) == {"Actual Positive", "Actual Negative"}
+    assert set(result.confusion.index) == {"Actual Positive", "Actual Negative"}
+    assert set(result.confusion.columns) == {"Predicted Positive", "Predicted Negative"}
 
 
 def test_true_blocks_validation_errors(small_named_txt_data):
@@ -328,8 +328,8 @@ def test_eval_metrics_structure(small_named_txt_data):
     assert all(0 <= val <= 1 for val in eval_result.metrics.to_numpy())
 
     assert eval_result.confusion.shape == (2, 2)
-    assert list(eval_result.confusion.index) == ["Predicted Negative", "Predicted Positive"]
-    assert list(eval_result.confusion.columns) == ["Actual Negative", "Actual Positive"]
+    assert list(eval_result.confusion.index) == ["Actual Positive", "Actual Negative"]
+    assert list(eval_result.confusion.columns) == ["Predicted Positive", "Predicted Negative"]
     assert (eval_result.confusion >= 0).all().all()
 
 
