+++ docs/user_guide/configuration_tuning.md
@@ -80,17 +80,17 @@ control_ann = {
 ```python
 control_ann = {
     'faiss': {
-        'index_type': ['flat', 'hnsw', 'lsh'], # Index type
-        'distance': str, # Distance metric
-        'k_search': int, # Number of neighbors to search
-        'path': str,     # Optional path to save index
+        'index_type': ['flat', 'hnsw', 'lsh'], # Index type (default: 'hnsw')
+        'distance': 'cosine', # Distance metric
+        'k_search': 30, # Number of neighbors to search
+        'path': None,     # Optional path to save index
 
-        'hnsw_M': int,               # Number of connections per element
-        'hnsw_ef_construction': int, # Size of dynamic candidate list (construction)
-        'hnsw_ef_search': int,       # Size of dynamic candidate list (search)
+        'hnsw_M': 32,               # Number of connections per element
+        'hnsw_ef_construction': 200, # Size of dynamic candidate list (construction)
+        'hnsw_ef_search': 200,       # Size of dynamic candidate list (search)
 
-        'lsh_nbits': int,        # (gets multiplied by dimensions) Number of bits for LSH
-        'lsh_rotate_data': bool, # Rotate data for LSH
+        'lsh_nbits': 2,        # (gets multiplied by dimensions) Number of bits for LSH
+        'lsh_rotate_data': True, # Rotate data for LSH
     }
 }
 ```
