+++ docs/examples/record_linkage.md
@@ -131,27 +131,16 @@ print(cis['txt'].head())
 
 ## Perform record linkage
 
-Initialize blocker instance and perform blocking with `hnsw` algorithm, `cosine` distance and custom parameters:
+Initialize blocker instance and perform blocking with `hnsw` algorithm and default parameters:
 
 ```python
 blocker = Blocker()
-
-control_ann = {
-    "hnsw": {
-        'distance': "cosine",
-        'M': 40,
-        'ef_c': 500,
-        'ef_s': 500
-    }
-}
 
 rec_lin_result = blocker.block(
     x=census['txt'],
     y=cis['txt'],   
     ann='hnsw',    
     verbose=1,      
-    control_ann=control_ann, 
-    # control_txt=control_txt, # let's leave this as default
     random_seed=42
 )
 
@@ -169,14 +158,14 @@ print(rec_lin_result)
 
 # ========================================================
 # Blocking based on the hnsw method.
-# Number of blocks: 23994
+# Number of blocks: 23993
 # Number of columns used for blocking: 1072
 # Reduction ratio: 0.999961
 # ========================================================
 # Distribution of the size of the blocks:
 # Block Size | Number of Blocks
-        #  2 | 23390          
-        #  3 | 590            
+        #  2 | 23388          
+        #  3 | 591            
         #  4 | 13             
         #  5 | 1    
 
@@ -230,16 +219,16 @@ census['x'] = range(len(census))
 cis['y'] = range(len(cis))
 
 # Find true matches using person_id
-matches = pd.merge(
+true_blocks = pd.merge(
     left=census[['PERSON_ID', 'x']],
     right=cis[['PERSON_ID', 'y']],
     on='PERSON_ID'
 )
 
 # Add block numbers
-matches['block'] = range(len(matches))
+true_blocks['block'] = range(len(true_blocks))
 
-matches.shape
+true_blocks.shape
 # (24043, 4)
 ```
 Let's sample 1000 pairs for which we will evaluate:
@@ -250,88 +239,30 @@ matches = matches.sample(1000, random_state=42)
 Now we can evaluate the algorithm:
 
 ```python
-# Perform blocking with evaluation
-eval_result = blocker.block(
-    x=census['txt'],
-    y=cis['txt'],
-    true_blocks=matches[['x', 'y', 'block']],
-    verbose=1,
-    ann='voyager',  # Try a different algorithm
-    random_state=42
-)
-
-# ===== creating tokens =====
-# ===== starting search (voyager, x, y: 25343,24613, t: 1072) =====
-# ===== creating graph =====
-
-# alternatively we can use the `eval` method for separation:
-# result = blocker.block(
-#     x=census['txt'],
-#     y=cis['txt'],
-#     verbose=1,
-#     ann='voyager',
-#     random_state=42
-# )
-# eval_result = blocker.eval(
-#     blocking_result=result,
-#     true_blocks=matches[['x', 'y', 'block']]
-#)
-# The procedure in both cases stays the same.
-
-# Note: We recommend using eval() method when evaluating larger datasets 
-# since it allows you to set the batch size for currently evaluated record pairs.
+eval_result = blocker.eval(rec_lin_result, true_blocks[['x', 'y', 'block']])
 ```
 
-and print results with evaluation metrics:
+and print the evaluation metrics:
 
 ```python
-print(eval_result)
-# ========================================================
-# Blocking based on the voyager method.
-# Number of blocks: 23922
-# Number of columns used for blocking: 1072
-# Reduction ratio: 0.999961
-# ========================================================
-# Distribution of the size of the blocks:
-# Block Size | Number of Blocks
-#          2 | 23248          
-#          3 | 659            
-#          4 | 13             
[…]
