### 2025-04-22 4d8313b
fix typo in example
