### 2025-04-22 5c3b339
add model2vec to dep
