+++ pyproject.toml
@@ -44,6 +44,7 @@ voyager = "^2.0.9"
 faiss-cpu = "^1.9.0"
 pandas = ">=2.2.0,<3.0.0"
 mlpack = ">=4.4.0, <5.0.0"
+model2vec = "^0.4.1"
 
 
 [tool.poetry.group.dev.dependencies]
