### 2025-04-22 f176d84
added embedding support for blockingpy (no tests, docs yet)[skip ci]
