### 2025-04-23 353e4d2
tests for text transformer
