### 2025-04-23 8c51a95
docs update after embedding
