### 2025-04-23 6bd9125
fix tests on ubuntu 311
