### 2025-04-26 3020770
new add_block_column method
