### 2025-04-26 f43d310
tests and docs update for the new method
