### 2025-04-27 1eb4b89
embedding example in docs
