+++ docs/examples/index.md
@@ -7,5 +7,6 @@
 record_linkage
 deduplication
 deduplication_2
+embedding_dedup
 recordlinkage_integration
 splink_integration
