+++ docs/examples/embedding_dedup.md
@@ -1,0 +1,210 @@
+# Deduplication with Embeddings
+
+This tutorial demonstrates how to use the `BlockingPy` library for deduplication using embeddings instead of n-gram shingles. It is based on the [Deduplication No. 2 tutorial](https://blockingpy.readthedocs.io/en/latest/examples/deduplication_2.html), but adapted to showcase the use of embeddings.
+
+Once again, we will use the  `RLdata10000` dataset taken from [RecordLinkage](https://cran.r-project.org/package=RecordLinkage) R package developed by Murat Sariyar
+and Andreas Borg. It contains 10 000 records in total where some have been duplicated with randomly generated errors. There are 9000 original records and 1000 duplicates.
+
+## Data Preparation
+
+Let's install `blockingpy`:
+
+```bash
+pip install blockingpy
+```
+
+Import necessary packages and functions:
+
+```python
+import pandas as pd
+from blockingpy import Blocker
+from blockingpy.datasets import load_deduplication_data
+```
+
+Let's load the data and take a look at first 5 rows:
+
+```python
+data = load_deduplication_data()
+data.head()
+
+# 	fname_c1	fname_c2	lname_c1	lname_c2   by	bm	bd	id  true_id
+# 0	FRANK	    NaN	        MUELLER	    NaN	       1967	9	27	1	3606
+# 1	MARTIN	    NaN	        SCHWARZ	    NaN	       1967	2	17	2	2560
+# 2	HERBERT	    NaN	        ZIMMERMANN  NaN	       1961	11	6	3	3892
+# 3	HANS	    NaN	        SCHMITT	    NaN	       1945	8	14	4	329
+# 4	UWE	    NaN	        KELLER	    NaN	       2000	7	5	5	1994
+```
+
+Now we need to prepare the `txt` column:
+
+```python
+data = data.fillna('')
+data[['by', 'bm', 'bd']] = data[['by', 'bm', 'bd']].astype('str')
+data['txt'] = (
+    data["fname_c1"] +
+    data["fname_c2"] +
+    data['lname_c1'] +
+    data['lname_c2'] +
+    data['by'] +
+    data['bm'] +
+    data['bd']
+    )   
+data['txt'].head()
+
+# 0         FRANKMUELLER1967927
+# 1        MARTINSCHWARZ1967217
+# 2    HERBERTZIMMERMANN1961116
+# 3          HANSSCHMITT1945814
+# 4             UWEKELLER200075
+# Name: txt, dtype: object
+```
+
+## Basic Deduplication
+
+We'll now perform basic deduplication with `hnsw` algorithm, but instead of character-level n-grams, the text will be encoded into dense embeddings before approximate nearest neighbor search.
+
+```python
+blocker = Blocker()
+
+control_txt = {
+    "encoder": "embedding",
+    "embedding": {
+        "model": "minishlab/potion-base-32M",
+        # for other customization options see 
+        # configuration in User Guide
+    }
+}
+
+dedup_result = blocker.block(
+    x=data['txt'],
+    ann='hnsw',
+    verbose=1,
+    random_seed=42,
+    control_txt=control_txt,
+)
+```
+
+We can now take a look at the results: 
+```python
+print(dedup_result)
+
+# ========================================================
+# Blocking based on the hnsw method.
+# Number of blocks: 2425
+# Number of columns used for blocking: 512
+# Reduction ratio: 0.999505
+# ========================================================
+# Distribution of the size of the blocks:
+# Block Size | Number of Blocks
+#          2 | 796            
+#          3 | 548            
+#          4 | 328            
+#          5 | 251            
+#          6 | 153            
+#          7 | 108            
+#          8 | 70             
+#          9 | 54             
+#         10 | 34             
[…]
