### 2025-04-28 8e0e260
minor updates to docs and logging
