+++ docs/html/_sources/user_guide/configuration_tuning.md.txt
@@ -1,0 +1,255 @@
+(configuration_tuning)=
+# Configuration and Tuning
+
+## Overview
+
+BlockingPy provides two main configuration interfaces:
+
+- control_txt: Text processing parameters
+- control_ann: ANN algorithm parameters
+
+## Text Processing Configuration (`control_txt`)
+
+The `control_txt` dictionary controls how text data is processed before blocking:
+
+```python
+control_txt = {
+        "encoder": "shingle",
+        "embedding": {
+            "model": "minishlab/potion-base-8M",
+            "normalize": True,
+            "max_length": 512,
+            "emb_batch_size": 1024,
+            "show_progress_bar": False,
+            "use_multiprocessing": True,
+            "multiprocessing_threshold": 10000,
+        },
+        "shingle": {
+            "n_shingles": 2,
+            "lowercase": True,
+            "strip_non_alphanum": True,
+            "max_features": 5000,
+        },
+    }
+```
+### Parameter Details
+
+`n_shingles` (default: `2`)
+
+- Controls the size of character n-grams
+- Larger values capture more context but increase dimensionality
+- Common values: 2-4
+- Impact: Higher values more precise but slower
+
+
+`max_features` (default: `5000`)
+
+- Maximum number of features in the document-term matrix
+- Controls memory usage and processing speed
+- Higher values may improve accuracy but increase memory usage
+- Adjust based on your dataset size and available memory
+
+
+`lowercase` (default: `True`)
+
+- Whether to convert text to lowercase
+- Usually keep True for better matching
+- Set to False if case is meaningful for your data
+
+
+`strip_non_alphanum` (default: `True`)
+
+- Remove non-alphanumeric characters
+- Usually keep True for cleaner matching
+- Set to False if special characters are important
+
+NOTE: `control_txt` is used only if the input is `pd.Series` as the other options were already processed.
+
+## ANN Algorithm Configuration (`control_ann`)
+
+Each algorithm has its own set of parameters in the `control_ann` dictionary. Overall `control_ann` should be in the following structure:
+
+```python
+control_ann = {
+    "random_seed" : None, # Alternative to setting the seed directly in the `Blocker` 
+    "faiss" : {
+        # parameters here
+    },
+    "voyager" : {},
+    "annoy" : {},
+    "lsh" : {},
+    "kd" : {},
+    "hnsw": {},
+    # you can specify only the dict of the algorithm you are using
+
+    "algo" : "lsh" or "kd" # specify if using lsh or kd
+
+}
+```
+
+
+### FAISS Configuration
+```python
+control_ann = {
+    'faiss': {
+        'index_type': ['flat', 'hnsw', 'lsh'], # Index type (default: 'hnsw')
+        'distance': 'cosine', # Distance metric
+        'k_search': 30, # Number of neighbors to search
+        'path': None,     # Optional path to save index
+
+        'hnsw_M': 32,               # Number of connections per element
+        'hnsw_ef_construction': 200, # Size of dynamic candidate list (construction)
+        'hnsw_ef_search': 200,       # Size of dynamic candidate list (search)
+
+        'lsh_nbits': 2,        # (gets multiplied by dimensions) Number of bits for LSH
+        'lsh_rotate_data': True, # Rotate data for LSH
+    }
+}
+```
+
+**Supported distance metrics**:
+
+- `euclidean` 
+- `cosine` (default)
+- `inner_product`
+- `l1`
+- `manhattan`
+- `linf`
+- `canberra`
+- `bray_curtis`
+- `jensen_shannon`
+
+***NOTE*** : Distance metrics do not apply to `lsh` index type.
+
+For more information about `faiss` see [here](https://github.com/facebookresearch/faiss/wiki/MetricType-and-distances).
+
+## Voyager Configuration
+
+```python
[…]
