+++ docs/html/_sources/examples/index.md.txt
@@ -1,0 +1,12 @@
+(examples)=
+# Examples
+
+```{toctree}
+:maxdepth: 1
+
+record_linkage
+deduplication
+deduplication_2
+embedding_dedup
+recordlinkage_integration
+splink_integration
