### 2025-04-29 3f74e8e
optimized eval
