+++ blockingpy/blocker.py
@@ -3,7 +3,6 @@ Contains the main Blocker class for record linkage
 and deduplication blocking.
 """
 
-import itertools
 import logging
 from collections import OrderedDict
 from typing import Any
@@ -158,8 +157,7 @@ class Blocker:
         2. Sparse matrices (scipy.sparse.csr_matrix) as a Document-Term Matrix (DTM)
         3. Dense matrices (numpy.ndarray) as a Document-Term Matrix (DTM)
 
-        For evaluation of larger datasets, we recommend using the separate eval() method
-        since it allows you to set the batch size for evaluation.
+        Evaluation of larger datasets can be done separately using the `eval` method.
 
         For text data, additional preprocessing is performed using
         the parameters in control_txt.
@@ -294,111 +292,13 @@ class Blocker:
         x_df["block"] = x_df["query_g"].apply(lambda x: x_block[x] if x in x_block else None)
 
         if true_blocks is not None:
-            logger.info("===== evaluating =====")
-            total_tn = total_fp = total_fn = total_tp = 0
-            batch_size = 1000
-
             if not deduplication:
-                unique_tb_x = true_blocks["x"].unique()
-                unique_tb_y = true_blocks["y"].unique()
-
-                true_x_blocks = true_blocks[["x", "block"]].drop_duplicates()
-                true_y_blocks = true_blocks[["y", "block"]].drop_duplicates()
-                pred_x_blocks = x_df[["x", "block"]].drop_duplicates()
-                pred_y_blocks = x_df[["y", "block"]].drop_duplicates()
-
-                total_batches_x = (len(unique_tb_x) + batch_size - 1) // batch_size
-                total_batches_y = (len(unique_tb_y) + batch_size - 1) // batch_size
-                total_batches = total_batches_x * total_batches_y
-
-                for start_idx_x in range(0, len(unique_tb_x), batch_size):
-
-                    current_batch_x = (start_idx_x // batch_size) + 1
-
-                    sub_x = unique_tb_x[start_idx_x : start_idx_x + batch_size]
-
-                    for start_idx_y in range(0, len(unique_tb_y), batch_size):
-                        current_batch_y = (start_idx_y // batch_size) + 1
-                        current_batch = ((current_batch_x - 1) * total_batches_y) + current_batch_y
-                        logger.info(f"Evaluating batch {current_batch} of {total_batches}")
-
-                        sub_y = unique_tb_y[start_idx_y : start_idx_y + batch_size]
-
-                        tp, fp, fn = self._eval_rl_batch(
-                            sub_x, sub_y, true_x_blocks, true_y_blocks, pred_x_blocks, pred_y_blocks
-                        )
-
-                        total_tp += tp
-                        total_fp += fp
-                        total_fn += fn
-                total_tn = len(unique_tb_x) * len(unique_tb_y) - total_tp - total_fp - total_fn
-
+                TP, FP, FN, TN = self._eval_rl(x_df, true_blocks)
             else:
-                x_df_long = (
-                    x_df.melt(id_vars=["block"], value_vars=["x", "y"], value_name="x_x")
-                    .drop_duplicates(subset=["x_x"])[["x_x", "block"]]
-                    .rename(columns={"x_x": "x"})
-                )
-                unique_tb_x = true_blocks["x"].unique()
+                TP, FP, FN, TN = self._eval_dedup(x_df, true_blocks)
 
-                total_batches_x = (len(unique_tb_x) + batch_size - 1) // batch_size
-                total_batches = total_batches_x * total_batches_x
-
[…]
