### 2025-04-29 7dfd938
fix faiss blocker
