### 2025-04-30 9938fe6
added controls_txt validation
