+++ docs/html/_sources/blocklib_comp.md.txt
@@ -1,35 +1,37 @@
 (blocklib_comp)=
 # BlockingPy vs blocklib - comparison
 
-Below we compare BlockingPy with blocklib, a similar library for blocking. We present results obtained by running algorithms from both libraries on 3 generated datasets. The datasets were generated using the `geco3` tool, which allows for controlled generation of datasets with duplicates. The datasets  resemble real-world personal information data with the fields such as name, 2nd name, surname, 2nd surname, dob, municipality, and country of origin. There are 1k, 10k and 100k records respectively, with 500, 5k and 50k duplicates in each dataset. For each original record, there are 0, 1, or 2 duplicates.
+Below we compare BlockingPy with blocklib, a similar library for blocking. We present results obtained by running algorithms from both libraries on 3 generated datasets. The datasets were generated using the `geco3` tool, which allows for controlled generation of datasets with duplicates. The datasets  resemble real-world personal information data with the fields such as name, 2nd name, surname, 2nd surname, dob, municipality, and country of origin. There are 1k, 10k and 100k records respectively, with 500, 5k and 50k duplicates in each dataset. For each original record, there are 0, 1, or 2 duplicates. The datasets and code to reproduce the results can be found [here](https://github.com/ncn-foreigners/BlockingPy/tree/main/benchmark). The results were obtained on 6 cores Intel i5 CPU with 16GB RAM.
 
-| Dataset size | Algorithm                     | Recall | Reduction Ratio | Blocking time (s) |
-|--------------|------------------------------|:------:|:---------------:|:-----------------:|
-| 1 500        | P-Sig (blocklib)             | 0.459 | 0.996 | 0.36 |
-| 1 500        | λ-fold LSH (blocklib)        | 0.427 | 0.993 | 0.40 |
-| 1 500        | **BlockingPy (faiss_hnsw)**  | **0.960** | **0.998** | 2.16 |
-| 1 500        | **BlockingPy (faiss_lsh)**   | **0.961** | **0.997** | **0.53** |
-| 1 500        | **BlockingPy (voyager)**     | **0.954** | **0.997** | 1.20 |
-| 15 000       | P-Sig (blocklib)             | 0.452 | 0.997 | 0.64 |
-| 15 000       | λ-fold LSH (blocklib)        | 0.420 | 0.994 | 2.94 |
-| 15 000       | **BlockingPy (faiss_hnsw)**  | **0.913** | **0.999** | 51.70 |
-| 15 000       | **BlockingPy (faiss_lsh)**   | **0.901** | **0.999** | **11.34** |
-| 15 000       | **BlockingPy (voyager)**     | **0.895** | **0.999** | 28.76 |
-| 150 000      | P-Sig (blocklib)             | 0.450 | 0.997 | 6.93 |
-| 150 000      | λ-fold LSH (blocklib)        | 0.413 | 0.994 | 32.03 |
-| 150 000      | **BlockingPy (faiss_hnsw)**  | **0.836** | **0.99996** | 1 020.05 |
-| 150 000      | **BlockingPy (faiss_lsh)**   | **0.818** | **0.99996** | **865.63** |
-| 150 000      | **BlockingPy (voyager)**     | **0.762** | **0.99995** | 648.07 |
+
+| algorithm                   | dataset\_size | time\_sec |   recall | reduction\_ratio | pairs (M) |
+| --------------------------- | ------------: | --------: | -------: | ---------------: | --------: |
+| P-Sig                       |         1 500 |     0.067 | 0.459168 |         0.996371 |  0.004080 |
+| λ-fold LSH                  |         1 500 |     0.210 | 0.426810 |         0.993112 |  0.007744 |
[…]
