### 2025-07-30 #232 @BERENZ
Hi, I can confirm that we are still interested!
