### 2025-08-04 #232 @crhea93
## Editor response to review:

---

## Editor comments
:wave: Hi @teald and @eliotwrobson  Thank you for volunteering to review
for pyOpenSci! I look forward to working with you all on this!

Please let me know when you have completed this :D

Please get in touch with any questions or concerns! Your review is due: August 31st, 2025
---

Reviewers: @teald @eliotwrobson
Due date: 08/31/2025
