### 2025-09-01 #232 @eliotwrobson
Apologies for the delay in reviewing (life has been extremely busy lately), but I managed to get through the items that didn't involve installation and playing around with the examples. I'll be able to get to this tomorrow.

EDIT:

I was able to download and play around with the package and have completed my review. I think this is a really useful package for anyone working with large datasets, so naturally researchers are a huge target audience. With a few changes from the items highlighted in my review, I think the package's polish could be significantly improved, making this easier to use for the target audience. Feel free to ask any questions about the items I raised, and if any are too labor intensive, we can discuss whether any of these have to be blocking.

- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

---

- I think including poetry installation instructions in the README just adds clutter, the pip instructions are fine on their own.
- It looks like you're missing a more comprehensive overview of the functions / classes available in the package in the API documentation. This is especially helpful to understand the different functionality provided by the package, and is part of the items above.
- If your code is being linted, you should add the linting as part of the automated workflow that runs your tests (or a separate workflow, but still this should be checked there somehow). It also looks like your code has type annotations, so adding mypy to your CI workflow would improve your code quality and maintainability as well.
- You should add a link directly to the examples from the documentation in the README as part of the basic usage.
- It was a bit tricky to get tests to run locally since the test workflow on GitHub doesn't use a virtual environment through poetry. I would suggest switching to this or using uv for this purpose.
- I think the code coverage can be improved by covering the `blocking_result.py` file more, and excluding the gpu file from coverage (or using mocks to cover the code there).
- Not a super expert about best practices here, but you use the logger to emit warnings in a few places, but I think it would be better to just use the warnings module to do this.
- Your package includes hardcoded datasets that are mainly used for just examples, but this adds to the size of the package while including data that many users may not use or care about. What would be better is to have this example data hosted on the repo for the project, but not part of the package itself, then use the [Pooch](https://github.com/fatiando/pooch) library to automate downloading of the data files to make the examples work.
- Not a hard requirement, but as someone who isn't a super expert (but still somewhat familiar with this topic) it would be helpful to have some discussion somewhere about the different algorithms and performance metrics. It looks like these are discussed in the associated paper, so it would be awesome if some of that discussion could be added to the documentation site (even if truncated).
