### 2025-09-01 #232 @BERENZ
Hi @eliotwrobson, thanks! Please note that a new version of the package arrived a couple of days ago and includes performance improvements as well as a new version that leverages GPUs for blocking (see the `blockingpy-gpu` version)."
