### 2025-09-09 #232 @akritaag
I've filled out the pre-review survey and here is my package review as well -

- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

---

1. This document was linked for a GPU version but returns a 404 error: https://blockingpy.readthedocs.io/en/latest/gpu/index.html
2. Issue: The package defaults to using FAISS (ann="faiss") but FAISS is not included as a required dependency in pyproject.toml. Impact: Users get ModuleNotFoundError: No module named 'faiss' when trying to use the default functionality. I have created an issue and a pull request for this.
3. Should there be examples to run other algorithms besides faiss? And scenarios where we might need other algorithms?
