### 2025-09-09 #232 @T-Strojny
Hello, thank you very much for your reviews. I just wanted to let you know that I will be able to address both of them next week.
