### 2025-09-09 #232 @akritaag
@T-Strojny I have a PR for the 2nd: https://github.com/ncn-foreigners/BlockingPy/issues/9
