### 2025-09-29 #232 @T-Strojny
Hello, here are the changes we've made:

Regarding @eliotwrobson’s review:

- Removed Poetry installation instructions from the README.
- Expanded and fixed the API section in docs.
- Linting, formatting, type checking workflow was added (ruff, mypy).
- Added a link to examples in the README.
- We've added venv using uv to the workflows.
- Coverage was improved by adding mocks for GPU file and expanding tests for `blocking_result.py` (60% -> 83%)
- logger warnings were replaced with `warnings` module
- Datasets are now fetched from data release on github with `pooch`
- Regarding metrics: we discuss them [here](https://blockingpy.readthedocs.io/en/latest/user_guide/evaluation_metrics.html), and there is also some information about speed and performance [here](https://blockingpy.readthedocs.io/en/latest/blocklib_comp.html). If this is not sufficient or you had something else in mind, I’d be happy to expand the documentation further.

As for @akritaag's review:

- the 404 error was fixed.
- I have added instructions to `CONTRIBUTING.md` on how to install the packages as editables to ensure everything works correctly. This should make it easier to test, develop, and contribute to the packages in the future. I also fixed some issues in the `pyproject.toml` files that previously caused problems during editable installs. Everything should work now if done according to the instructions.
- The documentation already includes examples using other algorithms such as `hnsw` and `voyager`. If this was not what you meant, could you clarify? I’ll be happy to address it.

Once again, thank you very much for your reviews.
