### 2025-10-01 #232 @crhea93
---

🎉 BlockingPy has been approved by pyOpenSci! Thank you @T-Strojny  for submitting BlockingPy and many thanks to @akritaag and @eliotwrobson  for reviewing this package! 😸

- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

---
