### 2026-03-09 #232 @BERENZ
Hi @crhea93 @isabelizimm I'm writing to let you know that the paper on the BlockingPy was just published in SoftwareX https://www.sciencedirect.com/science/article/pii/S2352711026000774. In the Acknowledgements we mention that software was reviewed and accepted to the PyOpenSci. Thank you once again for comments!

Edit: maybe we could close this issue?
