### 2026-04-13 #232 @eliotwrobson
@BERENZ sorry for taking a while to close this issue! We're working on some policy revisions right now so it's been a bit of a busy time. As part of this, we want to more prominently feature publications related to accepted packages, so if no objection we'd love to list the software X publication as part of the package information on our site.
